requirements.txt
litellm==1.82.0
# openai-agents  # optional: required for examples/agentic_vectorless_rag_demo.py
pymupdf==1.26.4
PyPDF2==3.0.1
python-dotenv==1.1.0
pyyaml==6.0.2
agentops>=0.4.0
fastapi
uvicorn
httpx
pydantic
langgraph
langchain-core
langchain-google-genai
python-dotenv
slack_sdk